diabetic_retinopathy_app/requirements.txt
Flask==2.0.1
numpy==1.21.2
tensorflow==2.3.0
tensorflow-estimator==2.3.0
Keras==2.4.3
Pillow==8.3.1
protobuf==3.20.1